pyproject.toml
[project]
name = "accessibility-audit-bot"
version = "0.1.0"
description = "FastAPI service combining axe-core scan results with LLM-prioritized remediation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "playwright>=1.46.0",
  "tavily-python>=0.5.0",
  "google-generativeai>=0.8.0",
  "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
